Adjusted weld-integration to build and run properly under weld 1.1.7 through 1.1.9. (refs JBAPP6-1659)
The first fix was to return a valid value from our ObserverMethod.getTransactionPhase() implementations.

The second fix was to eliminate ConversationalEvent[Impl].fire() because its implementation uses a method
that was removed in Weld 1.1.7. It appears that ConversationalEvent is now an unused class, so I will
create a Jira to investigate removing it in Errai 3.0.